[refactor] InputManager의 method를 각 Manager로 분리 및 Default Method를 통한 코드 최적화
