+++ EmployeeManager/src/main/java/DelInputManager.java
@@ -13,11 +13,6 @@ public class DelInputManager implements InputManagerInterface {
     }
 
     @Override
-    public List<String> getInfos() {
-        return null;
-    }
-
-    @Override
     public List<String> getOptions() {
         return options;
     }
@@ -31,15 +26,4 @@ public class DelInputManager implements InputManagerInterface {
     public String getValue() {
         return value;
     }
-
-    @Override
-    public String getChgKey() {
-        return null;
-    }
-
-    @Override
-    public String getChgValue() {
-        return null;
-    }
-
 }
