+++ EmployeeManager/src/main/java/SchInputManager.java
@@ -13,30 +13,17 @@ public class SchInputManager implements InputManagerInterface {
     }
 
     @Override
-    public List<String> getInfos() {
-        return null;
-    }
-
-    @Override
     public List<String> getOptions() {
         return options;
     }
+
     @Override
     public String getKey() {
         return key;
     }
+
     @Override
     public String getValue() {
         return value;
-    }
-
-    @Override
-    public String getChgKey() {
-        return null;
-    }
-
-    @Override
-    public String getChgValue() {
-        return null;
     }
 }
